+++ app/views/permission/index.php
@@ -53,7 +53,7 @@ define('SCREEN', 'perm');
                                 echo '<tr class="gradeX">';
                                 echo '<td class="text-center">' . _h($v['Key']) . '</td>';
                                 echo '<td class="text-center">' . _h($v['Name']) . '</td>';
-                                echo '<td class="text-center"><a href="' . get_base_url() . 'permission/' . _h($v['ID']) . '/" data-toggle="tooltip" data-placement="top" title="View/Edit" class="btn bg-yellow"><i class="fa fa-edit"></i></a></td>';
+                                echo '<td class="text-center"><a href="' . get_base_url() . 'permission/' . (int)_h($v['ID']) . '/" data-toggle="tooltip" data-placement="top" title="View/Edit" class="btn bg-yellow"><i class="fa fa-edit"></i></a></td>';
                                 echo '</tr>' . "\n";
                             }
                         }
